\name{PCGSE-package}
\alias{PCGSE-package}
\docType{package}
\title{
Implementation of the Principal Component Gene Set Enrichment (PCGSE) algorithm
}
\description{
Contains logic to compute the statistical association between gene sets and the principal components of 
experimental data using a two-stage competitive test.
}
\details{
\tabular{ll}{
Package: \tab PCGSE\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2014-03\cr
License: \tab GPL-3 \cr
}
Principal component gene set enrichment is performed using the function \code{pcgse}.
}
\author{
H. Robert Frost
}
\references{
Frost, HR, Li, Z, Moore, JH. Principal component gene set enrichment (PCGSE). Submitted.
}
\note{
This work was supported by the National Institutes of Health R01 grants LM010098, LM011360, EY022300, GM103506 and GM103534.
}
\keyword{ package }

