\name{PCGSE-package}
\alias{PCGSE-package}
\docType{package}
\title{
Implementation of the Principal Component Gene Set Enrichment (PCGSE) and Spectral Gene Set Enrichment (SGSE) algorithms
}
\description{
Contains logic to compute the statistical association between gene sets and the principal components of 
experimental data.
}
\details{
\tabular{ll}{
Package: \tab PCGSE\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2014-03\cr
License: \tab GPL-3 \cr
}
Principal component gene set enrichment is performed using the function \code{pcgse}. 
Spectral gene set enrichment is performed using the function \code{sgse}.
}
\author{
H. Robert Frost
}
\references{
    \itemize{
    \item Frost, H. R., Li, Z., and Moore, J. H. (2014). Principal component gene set enrichment (PCGSE). ArXiv e-prints. arXiv:1403.5148.
    \item Frost, H. R., Li, Z., and Moore, J. H. (2014). Spectral gene set enrichment (SGSE). ArXiv e-prints.
    }
}
\note{
This work was supported by the National Institutes of Health R01 grants LM010098, LM011360, EY022300, GM103506 and GM103534.
}
\keyword{ package }

